+++ src/bda/plone/orders/interfaces.py
@@ -4,3 +4,8 @@ from zope.interface import Interface
 class IOrdersExtensionLayer(Interface):
     """Browser layer for bda.plone.orders
     """
+
+
+class IVendor(Interface):
+    """Provide Vendor relevant information.
+    """